+++ crates/teloxide/src/dispatching/dialogue/storage.rs
@@ -9,9 +9,6 @@ mod redis_storage;
 #[cfg(feature = "sqlite-storage")]
 mod sqlite_storage;
 
-#[cfg(feature = "rocksdb-storage")]
-mod rocksdb_storage;
-
 use futures::future::BoxFuture;
 use teloxide_core::types::ChatId;
 
@@ -27,9 +24,6 @@ use std::sync::Arc;
 
 #[cfg(feature = "sqlite-storage")]
 pub use sqlite_storage::{SqliteStorage, SqliteStorageError};
-
-#[cfg(feature = "rocksdb-storage")]
-pub use rocksdb_storage::{RocksDbStorage, RocksDbStorageError};
 
 /// A storage with an erased error type.
 pub type ErasedStorage<D> =